[WIP] Add support for rdb version 8
RDB save format 8 includes 3 new record types, ZSET_2, MODULE, and
MODULE_2. This implements support for ZSET_2 while stubbing support for
MODULE_2. The original MODULE format isn't parsable without the module
to drive the parser, and this panics in that case.

Additionally, 64-bit integers are now supported as part of the length
encoding. This required a lot of changes from u32 to u64 across the
codebase. These changes are consistent with the Redis rdb.c
implementation which also works in terms of unsigned 64-bit integers.